+++ interfaces/Portal/src/app/program/program-team/program-team-popup/popup-content/change-team-member-popup-content/change-team-member-popup-content.component.ts
@@ -27,11 +27,16 @@ import { WarningLabelComponent } from './../../../../../shared/warning-label/war
 export class ChangeTeamMemberPopupContentComponent implements OnInit {
   @Input()
   public programId: number;
+
   @Input()
   public teamMemberRow: TeamMemberRow;
 
+  @Input()
+  public enableScope: boolean;
+
   private userId: number;
 
+  public scope = '';
   public searchQuery = '';
 
   public rolesList: Role[] = [];
@@ -48,6 +53,7 @@ export class ChangeTeamMemberPopupContentComponent implements OnInit {
     this.userId = this.teamMemberRow.id;
     this.getRoles();
     this.initiateSelectedRoleNames();
+    this.scope = this.teamMemberRow.scope;
   }
 
   public rolesAreSelected(): boolean {
@@ -77,6 +83,7 @@ export class ChangeTeamMemberPopupContentComponent implements OnInit {
       this.programId,
       this.userId,
       this.selectedRoleNames,
+      this.scope,
     );
     this.closeModal();
     this.teamMemberService.successPopup(