+++ services/121-service/seed-data/mock/visa-card.data.ts
@@ -1,14 +1,14 @@
 import { FspName } from '../../src/fsp/enum/fsp-name.enum';
 import { CustomDataAttributes } from '../../src/registration/enum/custom-data-attributes';
+import { LanguageEnum } from '../../src/registration/enum/language.enum';
 
 export const programIdVisa = 3;
 export const paymentNrVisa = 1;
 export const amountVisa = 22;
 
-export const referenceIdVisa = '2982g82bdsf89sdsd';
 export const registrationVisa = {
-  referenceId: referenceIdVisa,
-  preferredLanguage: 'en',
+  referenceId: '2982g82bdsf89sdsd',
+  preferredLanguage: LanguageEnum.en,
   paymentAmountMultiplier: 1,
   firstName: 'Jane',
   lastName: 'Doe',