+++ services/121-service/src/registration/registrations.controller.ts
@@ -349,8 +349,11 @@ export class RegistrationsController {
     name: 'dryRun',
     required: false,
     type: 'boolean',
-    description:
-      'When this parameter is set to `true`, the function will simulate the execution of the process without actually making any changes, so no registration statuses will be updated or messages will be sent.  Instead it will return data on how many registrations this action can be applied to. If this parameter is not included or is set to `false`, the function will execute normally. In both cases the response will be the same.',
+    description: `
+      Only when set explicitly to "true", this will simulate (and NOT actually DO) the action.
+      Instead it will return how many PA this action can be applied to.
+      So no registration statuses will be updated or messages will be sent.
+      `,
   })
   @ApiQuery({
     name: 'limit',
@@ -619,8 +622,11 @@ export class RegistrationsController {
     name: 'dryRun',
     required: false,
     type: 'boolean',
-    description:
-      'When this parameter is set to `true`, the function will simulate the execution of the process without actually making any changes, so no registrations will be deleted. Instead it will return data on how many PAs this action can be applied to. If this parameter is not included or is set to `false`, the function will execute normally. In both cases the response will be the same.',
+    description: `
+      Only when set explicitly to "true", this will simulate (and NOT actually DO) the action.
+      Instead it will return how many PA this action can be applied to.
+      No registrations will be deleted.
+      `,
   })
   @ApiQuery({
     name: 'limit',
@@ -789,8 +795,11 @@ export class RegistrationsController {
     name: 'dryRun',
     required: false,
     type: 'boolean',
-    description:
-      'When this parameter is set to `true`, the function will simulate the execution of the process without actually making any changes, so no messages will be sent. Instead it will return data on how many PAs this action can be applied to. If this parameter is not included or is set to `false`, the function will execute normally. In both cases the response will be the same.',
+    description: `
+      Only when set explicitly to "true", this will simulate (and NOT actually DO) the action.
+      Instead it will return how many PA this action can be applied to.
+      No messages will be sent.
+      `,
   })
   @ApiQuery({
     name: 'limit',