+++ services/121-service/src/scripts/seed-multiple-nlrc.ts
@@ -1,10 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource } from 'typeorm';
 import instanceNLRC from '../../seed-data/instance/instance-nlrc.json';
-import messageTemplateLVV from '../../seed-data/message-template/message-template-nlrc-lvv.json';
 import messageTemplateOCW from '../../seed-data/message-template/message-template-nlrc-ocw.json';
 import messageTemplatePV from '../../seed-data/message-template/message-template-nlrc-pv.json';
-import programLVV from '../../seed-data/program/program-nlrc-lvv.json';
 import programOCW from '../../seed-data/program/program-nlrc-ocw.json';
 import programPV from '../../seed-data/program/program-nlrc-pv.json';
 import { MessageTemplateService } from '../notifications/message-template/message-template.service';
@@ -34,21 +32,12 @@ export class SeedMultipleNLRC implements InterfaceScript {
     // Technically multiple instances could be loaded, but that should not be done
     await this.seedHelper.addInstance(instanceNLRC);
 
-    // ************************
-    // ***** Program LVV *****
-    // ************************
-
-    // ***** CREATE PROGRAM *****
-    const programEntityLVV = await this.seedHelper.addProgram(programLVV);
-
-    // ***** CREATE MESSAGE TEMPLATES *****
-    await this.seedHelper.addMessageTemplates(
-      messageTemplateLVV,
-      programEntityLVV,
+    // ***** SET SEQUENCE *****
+    // This is to keep PV and OCW program ids on respectively 2 and 3
+    // This to prevent differences between our local and prod dbs so we are less prone to mistakes
+    await this.dataSource.query(
+      `ALTER SEQUENCE "121-service".program_id_seq RESTART WITH 2;`,
     );
-
-    // ***** ASSIGN AIDWORKER TO PROGRAM WITH ROLES *****
-    await this.seedHelper.addDefaultUsers(programEntityLVV, debugScopes);
 
     // ************************
     // ***** Program PV *****