+++ services/121-service/test/helpers/assert.helper.ts
@@ -1,8 +1,6 @@
 import { MessageTemplateEntity } from '../../src/notifications/message-template/message-template.entity';
 import { RegistrationStatusEnum } from '../../src/registration/enum/registration-status.enum';
 import { RegistrationEntity } from '../../src/registration/registration.entity';
-import { waitFor } from '../../src/utils/waitFor.helper';
-import { getTransactions } from './program.helper';
 
 export const assertArraysAreEqual = (
   actualArray: any[],
@@ -52,43 +50,6 @@ export const assertObjectsAreEqual = (
     }
   }
 };
-
-export async function waitForPaymentTransactionsToComplete(
-  programId: number,
-  paymentReferences: string[],
-  accessToken: string,
-  maxWaitTimeMs: number,
-): Promise<void> {
-  const startTime = Date.now();
-  let allTransactionsSuccessful = false;
-
-  while (Date.now() - startTime < maxWaitTimeMs && !allTransactionsSuccessful) {
-    // Get payment transactions
-    const paymentTransactions = await getTransactions(
-      programId,
-      null,
-      null,
-      accessToken,
-    );
-
-    // Check if all transactions have a status of "success"
-    allTransactionsSuccessful = paymentReferences.every((referenceId) => {
-      const transaction = paymentTransactions.body.find(
-        (txn) => txn.referenceId === referenceId,
-      );
-      return transaction && transaction.status === 'success';
-    });
-
-    // If not all transactions are successful, wait for a short interval before checking again
-    if (!allTransactionsSuccessful) {
-      await waitFor(1000); // Wait for 1 second (adjust as needed)
-    }
-  }
-
-  if (!allTransactionsSuccessful) {
-    throw new Error(`Timeout waiting for payment transactions to complete`);
-  }
-}
 
 export function processMessagePlaceholders(
   messageTemplates: MessageTemplateEntity[],